requirements.txt
asgiref==3.7.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.2
Flask==3.0.2
future==0.18.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.2.0
pypng==0.20220715.0
python-dotenv==1.0.1
python-wialon==1.2.3
qrcode==7.4.2
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
vininfo==1.8.0
Werkzeug==3.0.1
